펄린값을 처리한 값이 머티리얼 배열의 최대 크기를 넘어서서 문제가 발생하였던 것으로 추정되며, 이를 해결하였다.
