+++ Assets/Scenes/GameScene/NavMesh-Level.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7973ad7d2f1f22240ba46f4c8caf2cee
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 23800000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
