+++ Assets/Scripts/CharacterController.cs
@@ -49,6 +49,11 @@ public class CharacterController : MonoBehaviour
     // Update is called once per frame
     void Update()
     {
+        ProcessInput();
+    }
+
+    void ProcessInput()
+    {
         if (Input.GetMouseButtonDown(1))    // RMB
         {
             playerDestination = GetRaycastHitpoint();
@@ -111,9 +116,6 @@ public class CharacterController : MonoBehaviour
         }
 
     }
-
-    ///
-    ///
 
     void Stop()
     {
