+++ Assets/Scripts/LevelManager.cs
@@ -8,136 +8,11 @@ using Object = UnityEngine.Object;
 
 using System;
 using UnityEditor;
-
-public class LevelData
-{
-    public GameObject tileObject;
-    public GameObject obstacleObject;
-    public Material[] mats;
-    public NavMeshSurface surface;
-
-    public LevelData(GameObject _tileObject, GameObject _obstaclceObject, Material[] _mats, NavMeshSurface _surface)
-    {
-        tileObject = _tileObject;
-        obstacleObject = _obstaclceObject;
-        mats = _mats;
-        surface = _surface;
-    }
-}
+using UnityEngine.AI;
+using Unity.VisualScripting;
 
 public class LevelManager
 {
-    /// Internal Classese
-    public class Tile
-    {
-        public Tile(GameObject _tile)
-        {
-            x = _tile.transform.position.x;
-            y = _tile.transform.position.z;
-            tile = _tile;
-        }
-
-        public Tile(float _x, float _y, GameObject _tile)
-        {
-            x = _x;
-            y = _y;
-            tile = _tile;
-        }
-
-        public float x;
-        public float y;
-        public GameObject tile;
-        public float perlinValue;
-    }
-
-    public class Rect
-    {
-        public Rect(float _minX, float _maxX, float _minY, float _maxY)
-        {
-            minX = _minX;
-            maxX = _maxX;
-            minY = _minY;
-            maxY = _maxY;
-        }
-
-        public float minX = 0;
-        public float maxX = 0;
-        public float minY = 0;
-        public float maxY = 0;
-    }
-
-    public class TileList
-    {
-        public List<Tile> list;
-        List<Tile> listToRemove;
-
-        public TileList()
-        {
-            list = new List<Tile>();
-            listToRemove = new List<Tile>();
-        }
-
-        public Tile GetTile(int _x, int _y)
-        {
-            //해당 위치에 타일이 있는지 검사
-            Tile tTileFound = list.Find(t => (t.x == _x) && (t.y == _y));
-
-            if (tTileFound == null)
-            {
-                return null;
-            }
-            else
-            {
-                return tTileFound;
-            }
-        }
-
-        public void Add(Tile _tile)
-        {
-            list.Add(_tile);
-        }
-
-        public void RemoveAt(int _x, int _y)
-        {
-            Tile tTileFound = list.Find(t => (t.x == _x) && (t.y == _y));
-
-            list.Remove(tTileFound);
-        }
-
-        public void RemoveIf(Predicate<Tile> match)
-        {
-            listToRemove = list.FindAll(match);
-
-            foreach(Tile t in listToRemove)
-            {
-                Object.Destroy(t.tile);
-                list.Remove(t);
-            }
-        }
-
-        public List<Tile> GetList()
-        {
-            return list;
-        }
-        public void ApplyMaterials(Material[] _mats)
-        {
-            foreach (var t in list)
-            {
-                int tIndex = Mathf.FloorToInt(t.perlinValue * (5.0f));
-                t.tile.GetComponent<MeshRenderer>().material = _mats[tIndex];
-            }
-        }
-
-        public void Show()
-        {
-            foreach (var t in list)
-            {
-                t.tile.SetActive(true);
-            }
-        }
-    }
-
-
     /// Properties
     // external
     GameObject playerObject;
@@ -150,10 +25,10 @@ public class LevelManager
     TileList tiles = new TileList();
 
     float range = 20.0f;
-    float updateTime = 2.0f;
+    float updateTime = 0.1f;
 
-    Rect playerRect;
-    Rect mapRect;
+    public Rect playerRect;
+    public Rect mapRect;
 
     float perlinSeed = Random.Range(100.0f, 1000.0f); // 범위는 어떻게 정해야하나?
 
@@ -193,42 +68,24 @@ public class LevelManager
             tiles.Show();
 
             //Debug.Log(tileList.Count);
-
-            navMeshSurface.BuildNavMesh();
-
             yield return new WaitForSeconds(updateTime);
         }
     }
 
-    public IEnumerator CreateObstacles()
+    public void BuildNavMeshData()
     {
-        while (true)
+        navMeshSurface.BuildNavMesh();
+    }
+
+    public IEnumerator UpdateNavMeshData()
+    {
+        while(true)
         {
-            GameObject obstacleTile;
-
-
-            Vector3 obstaclePosition = new Vector3(
-                Random.Range(playerObject.transform.position.x - 7, playerObject.transform.position.x + 7),
-                0.5f,
-                Random.Range(playerObject.transform.position.z - 7, playerObject.transform.position.z + 7));
-
-            GameObject t = Object.Instantiate(obstacleObject, obstaclePosition, obstacleObject.transform.rotation);
-
-            float height = Mathf.PerlinNoise(Random.Range(0.0f, 10.0f), Random.Range(0.0f, 10.0f));
-            int h = Mathf.RoundToInt(Mathf.Lerp(1, 4, height));
-
-            for (int i = 0; i < h; i++)
-            {
-                obstacleTile = Object.Instantiate(t.transform.GetChild(0).gameObject, new Vector3(0.0f, i, 0.0f), t.transform.rotation);
-                obstacleTile.transform.SetParent(t.transform, false);
-
-            }
-
-            t.SetActive(true);
-
-            yield return new WaitForSeconds(3.0f);
+            navMeshSurface.UpdateNavMesh(navMeshSurface.navMeshData);
+            yield return new WaitForSeconds(updateTime);
         }
     }
+
 
     /////////////////////////////////////////////////////////////////////////
     /////////////////////////////////////////////////////////////////////////
@@ -273,7 +130,6 @@ public class LevelManager
 
     void AddToTileList()
     {
-
         // player Rect을 순회
         for (int i = Mathf.FloorToInt(playerRect.minY); i < Mathf.CeilToInt(playerRect.maxY); i++)
         {
